requirements.txt
# Trading Bot Dependencies
python-binance==1.0.19
ccxt==4.1.77
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
ta==0.10.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
python-telegram-bot==20.7
flask==2.3.3
yfinance==0.2.28
alpaca-trade-api==3.0.2
oandapyV20==0.7.2 